Add deduction guides and concept magic so that the unique_ptr and shared_ptr ctors for FluentRegisterTarget work without a bunch of manual casting.
First off, many thanks to Andrew P and elbeno on the C++ Denver Meetup discord for helping me figure out how to make this work.

Secondly, an explanation as to why this was a problem to begin with.
RTF is meant to be used where IRegisterTarget is subclassed by user's code and then those objects are used with FluentRegisterTarget:
    class SomeTargetSubclass : public RTF::IRegisterTarget<A, D> { ... }
    auto real_target = std::make_shared<SomeTargetSubclass>();
    auto fluent_target = RTF::FluentRegisterTarget{ real_target };
But this didn't compile before due to the FluentRegisterTarget ctor taking a shared_ptr<IRegisterTarget<A,D>> specifically.
Deduction guides were needed to make this compile, at least for the shared_ptr cases.

There is a further complication, that even with deduction guides, when unique_ptr's are used instead of shared_ptr's the ctor call became ambiguous.
This is because shared_ptr can be constructed from unique_ptr (but not the reverse, which is why shared_ptr was not affected).

To disambiguate, the FluentRegisterTarget ctor needed to take a unique_ptr<T> (at first, allowing *any* contained type).
This is then further constrained using std::derived_from to ensure that T is derived from the IRegisterTarget base class.

Finally, there is *again* ambiguity (between shared_ptr and unique_ptr), this time with OwnedOrViewedObject.
The solution there was to explicitly cast the whatever_ptr<T> to whatever_ptr<IRegisterTarget<A,D>> in the member initializer.
This removed the implicit conversion which was the cause of the ambiguity in the first place.