Add `name` member variable to IRegisterTarget.
After doing some implementation using this library, it became apparent that providing the name member variable in this base class was highly desirable.
I can't think of a reason why one would want the previous behavior, other than cases where instances are singletons, so having a runtime-dynamic name is unnecessary.
If it's really wanted, it's easy to change back.