Massive refactor of IFluentRegisterTargetInterposer into IFluentInterposer.
This refactor changes the Interposer interface to be type-safe: it is templated on AddressType and DataType.
The main motivation for this was to consolidate the various paths involving interposer operation in FluentRegisterTarget.
It also allows the interposer to distinguish when the operation targets an address vs a register vs a field.

The new interposer interface is very similar to the old one, except now operations are overloaded using structs from the RTF::Operations namespace.
These new structs encapsulate the parameters of each operation so that the interposer can process them how it wishes (vs just getting an opaque string before).

When these structs contain an address/register/field member, it is encapsulated in a typedef:
either AddressRegFieldType for operations where all 3 are valid
or AddressRegType for operations where only an address or register is valid.

The interposer must handle this variant.
However, when RTF_INTEROP_RMF is not defined, the variant holds only one alternative: AddressType.

It would be better for the alias to be just AddressType (and not std::variant<AddressType>)
as this would greatly simplify Interposer code when RTF_INTEROP_RMF is not defined.
However, it would make the code being added in the next commit much uglier.
So I decided to favor myself instead of someone not using RMF.  Sorrynotsorry.