Add "default" implementations of IFluentInterposer<AT,DT> that fall back to a "consolidated" interposer.
This "consolidated" interposer has an API nearly identical to the old IFluentRegisterTargetInterposer interface.
This is on purpose to ease transition and provide an easy option for those that don't need advanced (type-safe) interposer stuff.

There is a createDefaultInterposerAdapters() that instantiates all 16 type-safe interposers that defer to the default IConsolidatedFluentInterposer.
Therefore this default consolidated interposer must be set *before* the call to createDefaultInterposerAdapters().

Arguments passed to createDefaultInterposerAdapters() will be passed to the 16 ctors.
The function is also templated on the type of Interposer to create, with a default (that is implemented by RTF).